+++ plugin.php
@@ -1,6 +1,6 @@
 <?php
 $id = 'ppco';
-$version = '1.2.0';
+$version = '1.2.1';
 $ilias_min_version = '5.4.0';
 $ilias_max_version = '6.999';
 $responsible = 'Theodor Truffer';