+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
 ## [1.3.0]
-* allow selection of playlists
-* change default height/width to 256px/450px
+* feature: allow selection of playlists
+* feature: synchronize access to folders when adding new video
 
 ## [1.2.0]
 * ILIAS 6 support