+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.3.0]
 * feature: allow selection of playlists
 * feature: synchronize access to folders when adding new video
+* feature: set max width in % (responsive) instead of fixed pixel dimensions
 
 ## [1.2.0]
 * ILIAS 6 support